panoptes/requirements.txt
annotated-types==0.7.0
fastapi==0.115.8
moviepy==2.1.2
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.0.1
ollama==0.4.7
pytest==8.3.4
python-multipart==0.0.20
pyjwt==2.10.1
passlib[bcrypt]
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
opencv-python==4.11.0.86
numpy
# pydantic==2.10.6
# pydantic_core==2.28.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
